service/requirements.txt
coloredlogs
Flask
flask_restx
transformers
python-dotenv
sentencepiece
protobuf
einops
cython
pyyaml
farm-haystack[inference]
ruamel.yaml